Invalid names/ids no longer make the script exit + small renamings.
In order to reprompt the user right when they enter an invalid name, ids are checked during the name input function, which has been renamed to read_names_and_ids. In addition to checking for invalid ids, the function now also checks if the same hero was entered twice.

get_hero_id now returns None instead of -1 if the id wasn't found.

hero_info was renamed to HERO_DATA, and has been made global so that it doesn't have to constantly be passed around to different functions. This will probably change again when the program is eventually restructured into some kind of class.

cooldowns_file was rename to HERO_DATA_FILE.